+++ src/OpenSRS/Helper/OpenSrsApi.php
@@ -103,7 +103,7 @@ class OpenSrsApi
             foreach ($nameServers as $i => $ns) {
                 $result['ns' . ($i + 1)] = [
                     'host' => $ns['name'],
-                    'ip' => $ns['ipaddress'] // No IP address available
+                    // 'ip' => $ns['ipaddress'] // No IP address available
                 ];
             }
         }
@@ -123,19 +123,23 @@ class OpenSrsApi
 
         $rawContactData = $rawContactData[$type];
 
-        return ContactData::create([
+        return ContactData::create(array_map(fn ($data) => empty($data) ? null : $data, [
             // 'id' => $type,
-            'name' => sprintf('%s %s', (string) $rawContactData['first_name'], (string) $rawContactData['last_name']),
-            'email' => (string) $rawContactData['email'],
-            'phone' => (string) $rawContactData['phone'],
-            'organisation' => (string) $rawContactData['org_name'],
-            'address1' => (string) $rawContactData['address1'],
-            'city' => (string) $rawContactData['city'],
-            'state' => $rawContactData['state'],
-            'postcode' => (string) $rawContactData['postal_code'],
-            'country_code' => (string) $rawContactData['country'],
+            'name' => trim(sprintf(
+                '%s %s',
+                $rawContactData['first_name'] ?? null,
+                $rawContactData['last_name'] ?? null
+            )),
+            'email' => strval($rawContactData['email'] ?? null),
+            'phone' => strval($rawContactData['phone'] ?? null),
+            'organisation' => strval($rawContactData['org_name'] ?? null),
+            'address1' => strval($rawContactData['address1'] ?? null),
+            'city' => strval($rawContactData['city'] ?? null),
+            'state' => strval($rawContactData['state'] ?? null),
+            'postcode' => strval($rawContactData['postal_code'] ?? null),
+            'country_code' => strval($rawContactData['country'] ?? null),
             'type' => $type,
-        ]);
+        ]));
     }
 
     /**
@@ -173,7 +177,7 @@ class OpenSrsApi
             self::XML_INDENT . '</body>' . self::CRLF .
             '</OPS_envelope>';
 
-        $response = $this->client->request('POST', self::getApiEndpoint(!!$this->configuration->debug), [
+        $response = $this->client->request('POST', self::getApiEndpoint(!!$this->configuration->sandbox), [
             'body' => $xml,
             'headers' => [
                 'User-Agent' => 'Upmind/ProvisionProviders/DomainNames/OpenSRS',
@@ -331,7 +335,13 @@ class OpenSrsApi
         }
 
         if ((int)$data['is_success'] === 0 && !in_array($data['response_code'], [200, 212])) {
-            throw static::errorResult('Registrar API Error: ' . $data['response_text'], $data);
+            $errorMessage = 'Registrar API Error: ' . $data['response_text'];
+
+            if ($data['response_code'] == 400) {
+                $errorMessage = 'Registrar API Authentication Error';
+            }
+
+            throw static::errorResult($errorMessage, $data);
         }
 
         return $data;