+++ src/OpenSRS/Provider.php
@@ -476,7 +476,10 @@ class Provider extends DomainNames implements ProviderInterface
                 ]
             ]);
         } catch (ProvisionFunctionError $e) {
-            if (Str::contains($e->getMessage(), 'Authentication Error')) {
+            if (
+                Str::contains($e->getMessage(), 'Authentication Error')
+                && !Str::contains($e->getMessage(), 'Registrar API Authentication Error')
+            ) {
                 // this actually means domain not found
                 $this->errorResult('Domain name not found', $e->getData(), $e->getDebug(), $e);
             }