+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All notable changes to the package will be documented in this file.
 
+## [v2.8.8](https://github.com/upmind-automation/provision-provider-domain-names/releases/tag/v2.8.8) - 2023-09-11
+
+- Update LogicBoxes error message formatting; fix truncation issue where input string contains a period
+- Update LogicBoxes API error results to include response in result data
+
+## [v2.8.7](https://github.com/upmind-automation/provision-provider-domain-names/releases/tag/v2.8.7) - 2023-08-29
+
+- Fix DacParams sld validation
+
+## [v2.8.6](https://github.com/upmind-automation/provision-provider-domain-names/releases/tag/v2.8.6) - 2023-08-23
+
+- Remove min length rules from RealtimeRegister configuration
+
+## [v2.8.5](https://github.com/upmind-automation/provision-provider-domain-names/releases/tag/v2.8.5) - 2023-08-01
+
+- Fix OpenSRS debug/sandbox requests
+- Improve OpenSRS improve auth error message
+- Fix OpenSRS GetInfo where contact values are missing
+- Fix OpenSRS empty string IPs returned for nameservers
+
 ## [v2.8.4](https://github.com/upmind-automation/provision-provider-domain-names/releases/tag/v2.8.4) - 2023-07-17
 
 - Update NameSilo register() to use Utils::tldSupportsWhoisPrivacy()